Move the renaming to the gradle
