+++ analytics-sidecar-client/src/test/java/org/apache/cassandra/sidecar/client/HttpClientConfigTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package client;
+package org.apache.cassandra.sidecar.client;
 
 import java.io.InputStream;
