+++ analytics-sidecar-client/src/test/java/org/apache/cassandra/sidecar/client/SidecarClientConfigTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package client;
+package org.apache.cassandra.sidecar.client;
 
 import org.junit.jupiter.api.Test;
