+++ analytics-sidecar-client/src/test/java/org/apache/cassandra/sidecar/client/selection/OrderedInstanceSelectionPolicyTest.java
@@ -16,41 +16,64 @@
  * limitations under the License.
  */
 
-package client.selection;
+package org.apache.cassandra.sidecar.client.selection;
 
+import java.util.Arrays;
 import java.util.Iterator;
+import java.util.List;
 import java.util.NoSuchElementException;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import org.apache.cassandra.sidecar.client.SidecarInstancesProvider;
+import org.apache.cassandra.sidecar.client.SimpleSidecarInstancesProvider;
 import org.apache.cassandra.sidecar.client.SidecarInstance;
 import org.apache.cassandra.sidecar.client.selection.InstanceSelectionPolicy;
-import org.apache.cassandra.sidecar.client.selection.SingleInstanceSelectionPolicy;
+import org.apache.cassandra.sidecar.client.selection.OrderedInstanceSelectionPolicy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.mock;
 
-class SingleInstanceSelectionPolicyTest
+/**
+ * Unit tests for the {@link OrderedInstanceSelectionPolicy}
+ */
+class OrderedInstanceSelectionPolicyTest
 {
-    SidecarInstance mockSidecarInstance;
+    List<SidecarInstance> mockInstanceList;
+    SidecarInstance mockInstance1;
+    SidecarInstance mockInstance2;
+    SidecarInstance mockInstance3;
+    SidecarInstance mockInstance4;
 
     @BeforeEach
     void setup()
     {
-        mockSidecarInstance = mock(SidecarInstance.class);
+        mockInstance1 = mock(SidecarInstance.class);
+        mockInstance2 = mock(SidecarInstance.class);
+        mockInstance3 = mock(SidecarInstance.class);
+        mockInstance4 = mock(SidecarInstance.class);
+        mockInstanceList = Arrays.asList(mockInstance1, mockInstance2, mockInstance3, mockInstance4);
     }
 
     @Test
-    void testIterator()
+    public void testIterator()
     {
-        InstanceSelectionPolicy instanceSelectionPolicy = new SingleInstanceSelectionPolicy(mockSidecarInstance);
+
+        SidecarInstancesProvider provider = new SimpleSidecarInstancesProvider(mockInstanceList);
+        InstanceSelectionPolicy instanceSelectionPolicy = new OrderedInstanceSelectionPolicy(provider);
         Iterator<SidecarInstance> iterator = instanceSelectionPolicy.iterator();
 
         assertThat(iterator.hasNext()).isTrue().as("Expected to be true");
         assertThat(iterator.hasNext()).isTrue().as("Test idempotency of hasNext by running it again");
-        assertThat(iterator.next()).isSameAs(mockSidecarInstance);
+        assertThat(iterator.next()).isSameAs(mockInstance1);
+        assertThat(iterator.hasNext()).isTrue();
+        assertThat(iterator.next()).isSameAs(mockInstance2);
+        assertThat(iterator.hasNext()).isTrue();
+        assertThat(iterator.next()).isSameAs(mockInstance3);
+        assertThat(iterator.hasNext()).isTrue();
+        assertThat(iterator.next()).isSameAs(mockInstance4);
         assertThat(iterator.hasNext()).isFalse().as("Expected to be false");
         assertThat(iterator.hasNext()).isFalse().as("Test idempotency of hasNext by running it again");
         assertThatExceptionOfType(NoSuchElementException.class).isThrownBy(iterator::next);
